refactor(layout.component.ts): Moved layout to the core directory and changed it into a standalone component
