+++ src/BurstChat.Client/src/app/components/core/layout/layout.component.scss
@@ -1,6 +1,6 @@
-@import '../../../../../scss/theme';
-@import '../../../../../scss/loader';
-@import '../../../../../scss/sizing';
+@import '../../../../scss/theme';
+@import '../../../../scss/loader';
+@import '../../../../scss/sizing';
 
 @-webkit-keyframes layout-fade-in {
     0% {
