+++ src/BurstChat.Client/src/app/components/core/layout/layout.component.spec.ts
