+++ src/BurstChat.Client/src/app/modules/burst/burst.routing.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { LayoutComponent } from 'src/app/modules/burst/components/layout/layout.component';
+import { LayoutComponent } from 'src/app/components/core/layout/layout.component';
 
 const routes: Routes = [
     {
