--- src/BurstChat.Client/src/app/modules/burst/components/layout/layout.component.html
@@ -1,21 +1,0 @@
-<div class="layout" *ngIf="!loading">
-
-    <burst-sidebar></burst-sidebar>
-
-    <main>
-
-      <router-outlet></router-outlet>
-
-    </main>
-
-    <burst-incoming-call></burst-incoming-call>
-
-    <burst-add-server></burst-add-server>
-
-</div>
-
-<div class="layout-loader" *ngIf="loading">
-
-    <div class="loader"></div>
-
-</div>
